refactor(http): use custom result type
