+++ gitolith-core/src/http/mod.rs
@@ -1,5 +1,2 @@
 pub mod router;
 pub mod handlers;
-
-
-pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
